chore: use latest harmony changes
