+++ src/extensions/paper/index.ts
@@ -1,5 +1,4 @@
-export { Paper } from './paper';
-export { default as PaperExt } from './paper.extension';
+export { PaperExtension } from './paper.extension';
 export { Command, CLIArgs, PaperOptions, Flags, GenericObject } from './command';
 export { getID } from './registry';
 export * from './exceptions';
