+++ src/extensions/paper/readme.md
@@ -1,10 +1,9 @@
 # Paper
-Paper is a framework for building modern command line applications using React components.
-It uses [Ink](https://github.com/vadimdemedes/ink) and [yargs](https://github.com/yargs/yargs) behind the scences to deliver a unified experience for building amazing component-based CLI apps.
+Paper is an extension for building modern terminal commands using React components.
+It uses [Ink](https://github.com/vadimdemedes/ink) and [commander](https://github.com/tj/commander.js/) behind the scences to deliver a unified experience for building amazing component-based CLI apps.
 
 ## Features
 - Allow to register new commands
-- Use yargs commands.
-- Render stdout with Ink.
-- Progress API
+- Use commander commands.
+- Render to stdout with Ink.
 - move to new command definition format (see example at `run.cmd.tx`)
